requirements.txt
python-dotenv
requests
websockets
pydantic
PyYAML
